Create notion of $ARBDOOMDIR for consistent sourcing of ps and kill
The arbdoom.py script will now use the ps and kill scripts from
$ARBDOOMDIR if present or the -m/--arbdoomdir directory. This was done
to allow for the installation of arbdoom. Since it's annoying to have to
specify the $ARBDOOMDIR before running arbdoom, the arbdoom wrapper
script was created so that running arbdoom is as simple as ./arbdoom. I
should note that it may be desireable to have the ARBCONFIG and ARBDIR
values set in there.

Also: I think the ARB* environment variable changes that are used in
Arbiter's tools/ haven't yet been released publicly in yet...